Archive latest version before switching to mongodb
Up to this point, I was working on a tree traverser that scraped all the
different attributes possible in an oid's description table to build out
a relational database table.  This proved unreasonable and I have since
moved to mongodb which supports arbitrary field membership per document.

This version also includes the results of the time I got lucky and
managed to load 1.3.6.1.4.1.html.  It seems to be much too large for
oidref to generate and very often times out before loading.  From the
archived webpage, I converted the oid's children table into a json file
which will dramatically speed up processing of its sub tree.