Add basic "detailed data" scraper
I decided to store the detailed descriptions in markdown because it is
easy to convert to and from html.

However, around every corner is another curveball.  I figured this
section would be easy, but it turns out a lot (if not all) of the
registration authority sections have unclosed paragraph tags meaning the
paragraphs encompass all subsequent sections.  This prevents the capture
of the registraion authority data by the scraper.