Prepare for scraping the children of enterprise
There were a couple big changes made here.  The first being that
entrypoint now loads convert_enterprise/enterprise_compact.json in order
to get the list of enterprise's children to scrape from.  From there
it calls traverse_tree on each of them as if they are a root node.

In order to facilitate resuming the scrape in the event of a crash
(which seems inevitable with the past runs), sys.eventhook has been
replaced to print the root node (child of enterprise) currently being
scraped (it also calls the standard excepthook afterwards).  This
should help when using the drill_down branch later by skipping root
nodes that have already been scraped.