Fix unclosed paragraph truncation
The solution I found to the __str__ problem in the last commit was to
recreate the parse tree under (and including) the registration authority
Tag using a dft until the first h3 tag is found.  The h3 tag isn't
included and the stack is unwound.  Running str() on this truncated tree
is much faster than including the children/brother tables and everything
under the unclosed paragraph.